Move SizeTree to top-level package
